requirements.txt
alembic==1.4.2
appdirs==1.4.4
appnope==0.1.0
autopep8==1.4.4
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
bs4==0.0.1
bwsi-grader==1.8.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colored==1.4.2
decorator==4.4.2
distlib==0.3.1
dnspython==2.0.0
email-validator==1.1.1
filelock==3.0.12
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.4
Flask-SSLify==0.1.5
flask-talisman==0.7.0
Flask-WTF==0.14.3
gcloud==0.18.3
googleapis-common-protos==1.52.0
gunicorn==20.0.4
heroku==0.1.4
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
ipython==7.15.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.16.0
jsonify==0.5
lxml==4.4.2
Mako==1.1.3
MarkupSafe==1.1.1
nltk==3.5
numpy==1.18.3
oauth2client==4.1.3
parso==0.7.0
passlib==1.7.2
pexpect==4.8.0
pickleshare==0.7.5
plivo==4.9.0
prompt-toolkit==3.0.5
protobuf==4.0.0rc2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==1.5
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
regex==2020.7.14
requests==2.24.0
rsa==4.6
scapy==2.4.3
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
stripe==2.48.0
tqdm==4.48.0
traitlets==4.3.3
twilio==6.44.0
urllib3==1.25.9
virtualenv==20.0.28
wcwidth==0.2.4
Werkzeug==1.0.1
wget==3.2
WTForms==2.3.1
zipp==3.1.0
